requirements.txt
streamlit==1.38.0
pandas==2.2.2
plotly==5.23.0
python-dotenv==1.0.1
google-analytics-data==0.18.6
google-auth==2.34.0
openai==1.43.0
